pyproject.toml
[project]
name = "kaggle-join-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "unonao", email = "naokiuno634@gmail.com" }
]
dependencies = [
    "beautifulsoup4>=4.12.3",
    "pip>=24.0",
    "selenium>=4.19.0",
    "kaggle>=1.6.8",
    "numpy>=1.26.4",
    "gspread>=6.1.0",
    "google-auth>=2.29.0",
    "ruff>=0.3.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "jupyterlab>=4.1.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kaggle_join_bot"]
